Fix integer overflow in expression (-INT64_MIN)

+++ src/cc_bstring.c
@@ -140,7 +140,7 @@ bstring_atoi64(int64_t *i64, struct bstring *str)
         // overflow check
         if (offset == CC_INT64_MAXLEN - 2) {
             if (sign < 0 && *i64 == INT64_MIN / 10 &&
-                    c - '0' > (uint64_t)(-INT64_MIN) % 10) {
+                    c - '0' > -(INT64_MIN % 10)) {
                 return CC_ERROR;
             }
             if (sign > 0 && *i64 == INT64_MAX / 10 &&
